🎈 feat(mobile/app/auth): restructure auth navigation and add onboarding page
