+++ apps/mobile/src/app/(auth)/_layout.tsx
@@ -1,8 +1,12 @@
 import { Stack } from "expo-router";
 import { ThemeProvider } from "../../layers/app/providers";
-import { useNavigationGuard } from "../../layers/features/navigation-guard";
+import { useNavigationGuard } from "../../layers/features/auth";
 import { LoginPageHeader } from "../../layers/pages/login";
 import { WaveHeader } from "../../layers/widgets/layout";
+
+export const unstable_settings = {
+	initialRouteName: "onboarding",
+};
 
 const AuthLayout = () => {
 	const navigationGuard = useNavigationGuard();
@@ -15,14 +19,9 @@ const AuthLayout = () => {
 				}}
 			>
 				<Stack.Protected guard={navigationGuard.data === "unauthenticated"}>
-					<Stack.Screen name="index" options={{ headerShown: false }} />
+					<Stack.Screen name="onboarding" options={{ headerShown: false }} />
 					<Stack.Screen name="login" options={{ header: LoginPageHeader }} />
-					<Stack.Screen
-						name="signup"
-						options={{
-							header: () => <WaveHeader title="Sign Up" enableBackButton />,
-						}}
-					/>
+					<Stack.Screen name="signup" options={{ headerShown: false }} />
 					<Stack.Screen
 						name="forgot-password"
 						options={{ header: () => <WaveHeader title="Password Reset" enableBackButton /> }}
@@ -45,10 +44,6 @@ const AuthLayout = () => {
 
 				<Stack.Protected guard={navigationGuard.data === "personalize"}>
 					<Stack.Screen name="personalize" options={{ headerShown: false }} />
-				</Stack.Protected>
-
-				<Stack.Protected guard={navigationGuard.data === "ready"}>
-					<Stack.Screen name="ready" options={{ headerShown: false }} />
 				</Stack.Protected>
 			</Stack>
 		</ThemeProvider>
