+++ apps/mobile/src/app/(auth)/onboarding.tsx
