--- apps/mobile/src/app/(auth)/ready.tsx
@@ -1,11 +1,0 @@
-import { Text, View } from "react-native";
-
-const ReadyPage = () => {
-	return (
-		<View className="flex flex-1 items-center justify-center">
-			<Text>Ready</Text>
-		</View>
-	);
-};
-
-export default ReadyPage;
